+++ src/Web/Telegram/API/Bot/Responses.hs
@@ -28,6 +28,7 @@ module Web.Telegram.API.Bot.Responses
 
 import           Data.Aeson
 import           GHC.Generics
+import           Data.Int (Int64)
 import           Web.Telegram.API.Bot.Data
 import           Web.Telegram.API.Bot.JsonExt
 
@@ -39,8 +40,8 @@ data Response a = Response
 
 data ResponseParameters = ResponseParameters
   {
-    res_migrate_to_chat_id :: Maybe Int -- ^ The group has been migrated to a supergroup with the specified identifier. This number may be greater than 32 bits and some programming languages may have difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or double-precision float type are safe for storing this identifier.
-  , res_retry_after        :: Maybe Int -- ^ In case of exceeding flood control, the number of seconds left to wait before the request can be repeated
+    res_migrate_to_chat_id :: Maybe Int64 -- ^ The group has been migrated to a supergroup with the specified identifier. This number may be greater than 32 bits and some programming languages may have difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or double-precision float type are safe for storing this identifier.
+  , res_retry_after        :: Maybe Int   -- ^ In case of exceeding flood control, the number of seconds left to wait before the request can be repeated
   } deriving (Show, Generic)
 
 instance FromJSON ResponseParameters where
